leds: TI LMU: fix u8 variable comparisons with less than zero
The u8 variables ramp_ups and ramp_downs are being compared to less
than zero, this will always be false.  Fix this by making the ramp
variables ints.

Addresses-Coverity: ("Unsigned compared against 0")
Fixes: 9a8e66ebeaa2 ("leds: TI LMU: Add common code for TI LMU devices")
Reported-by: kbuild test robot <lkp@intel.com>
Reported-by: Dan Carpenter <dan.carpenter@oracle.com>

+++ drivers/leds/leds-ti-lmu-common.c
@@ -84,7 +84,7 @@ static int ti_lmu_common_convert_ramp_to_index(unsigned int usec)
 int ti_lmu_common_set_ramp(struct ti_lmu_bank *lmu_bank)
 {
 	struct regmap *regmap = lmu_bank->regmap;
-	u8 ramp, ramp_up, ramp_down;
+	int ramp, ramp_up, ramp_down;
 
 	if (lmu_bank->ramp_up_usec == 0 && lmu_bank->ramp_down_usec == 0) {
 		ramp_up = 0;
